requirements.txt
playwright~=1.58
pytest~=9.0
pytest-playwright~=0.7
pytest-base-url~=2.1
pytest-rerunfailures~=16.0
allure-pytest~=2.15
